refactor: relocate economy test imperative runners to consumers (Refs #4108)
