+++ packages/colonizethis_economy_test_support/lib/colonizethis_economy_test_support.dart
@@ -8,18 +8,14 @@
 /// dart` copies into one canonical location (Refs #3615 Cluster 6).
 library colonizethis_economy_test_support;
 export 'src/build_cost_scenarios.dart';
-export 'src/commodity_totals_expectations.dart';
 export 'src/commodity_totals_scenarios.dart';
 export 'src/boycott_blocked_commodities_scenarios.dart';
 export 'src/boycott_blocked_commodities_expectations.dart';
 export 'src/boycott_blocked_commodities_test_support.dart';
-export 'src/consumption_expectations.dart';
-export 'src/consumption_phases_expectations.dart';
 export 'src/consumption_phases_scenarios.dart';
 export 'src/consumption_scenarios.dart';
 export 'src/cost_check_scenarios.dart';
 export 'src/core_economy_test_support.dart';
-export 'src/economy_extraction_expectations.dart';
 export 'src/economy_extraction_scenarios.dart';
 export 'src/economy_riches_to_treasury_scenarios.dart';
 export 'src/economy_production_scenarios.dart';
@@ -38,7 +34,6 @@ export 'src/resource_extractor_expectations.dart';
 export 'src/resource_extractor_scenarios.dart';
 export 'src/scenario_runner.dart';
 export 'src/sea_transport_scenarios.dart';
-export 'src/tile_extraction_contribution_expectations.dart';
 export 'src/tile_extraction_contribution_scenarios.dart';
 export 'src/tile_extraction_pipeline_scenarios.dart';
 export 'src/trade_cargo_capacity_scenarios.dart';
@@ -53,7 +48,6 @@ export 'src/town_manufacturing_bonus_expectations.dart';
 export 'src/town_manufacturing_bonus_scenarios.dart';
 export 'src/trade_interception_scenarios.dart';
 export 'src/worker_action_cost_scenarios.dart';
-export 'src/worker_economy_expectations.dart';
 export 'src/worker_economy_scenarios.dart';
 export 'src/trade_order_admission_scenarios.dart';
 export 'src/trade_order_factory.dart';
