+++ packages/colonizethis_economy_test_support/lib/src/sea_transport_scenarios.dart
@@ -2,12 +2,15 @@
 // Table-driven sea transport scenarios (Refs #3939 phase 3, #3979).
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
-import 'sea_transport_expectations.dart';
+
+/// Pins for [cargoHoldsForHomeFleet] rows.
+typedef CargoHoldsPins = ({List<Fleet>? fleets, String playerId, int? expectedExact, bool expectNonNegativeOnly, bool parityFleetsById});
+
+CargoHoldsForHomeFleetScenario cargoHoldsScenario({required String label, required CargoHoldsPins pins}) => (label: label, pins: pins, refs: null);
+
 /// One row in [cargoHoldsForHomeFleetScenarios] (Refs #3979).
 typedef CargoHoldsForHomeFleetScenario = ({String label, CargoHoldsPins pins, String? refs});
-void runCargoHoldsForHomeFleetScenario(CargoHoldsForHomeFleetScenario scenario) {
-  runCargoHoldsExpectation(scenario.pins);
-}
+
 /// Canonical scenarios for [cargoHoldsForHomeFleet].
 List<CargoHoldsForHomeFleetScenario> cargoHoldsForHomeFleetScenarios() => [
   cargoHoldsScenario(label: 'returns 0 when no home fleet exists', pins: (fleets: null, playerId: 'p1', expectedExact: null, expectNonNegativeOnly: true, parityFleetsById: false)),
@@ -49,11 +52,15 @@ List<CargoHoldsForHomeFleetScenario> cargoHoldsForHomeFleetScenarios() => [
     ),
   ),
 ];
+
+/// Pins for [allocateOverseasToStockpile] rows.
+typedef AllocateOverseasPins = ({Map<String, int> overseas, int cargoHolds, List<CommodityCategory>? priorityOrder, bool expectEmpty, int? expectedTotal, Map<String, int?>? expectedDelivered});
+
+AllocateOverseasToStockpileScenario allocateOverseasScenario({required String label, required AllocateOverseasPins pins}) => (label: label, pins: pins, refs: null);
+
 /// One row in [allocateOverseasToStockpileScenarios] (Refs #3979).
 typedef AllocateOverseasToStockpileScenario = ({String label, AllocateOverseasPins pins, String? refs});
-void runAllocateOverseasToStockpileScenario(AllocateOverseasToStockpileScenario scenario) {
-  runAllocateOverseasExpectation(scenario.pins);
-}
+
 /// Canonical scenarios for [allocateOverseasToStockpile].
 List<AllocateOverseasToStockpileScenario> allocateOverseasToStockpileScenarios() => [
   allocateOverseasScenario(label: 'returns empty when overseas is empty', pins: (overseas: {}, cargoHolds: 10, priorityOrder: null, expectEmpty: true, expectedTotal: null, expectedDelivered: null)),
