+++ packages/colonizethis_economy_test_support/lib/src/tile_extraction_contribution_scenarios.dart
@@ -2,7 +2,17 @@
 // Table-driven `computeTileExtractionContributionForPlayer` scenarios (Refs #3939, #3979).
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'extraction_fixture_support.dart';
-import 'tile_extraction_contribution_expectations.dart';
+
+/// Pins for connected-tile extraction contribution rows.
+typedef TileContributionConnectedPin = ({String commodityId, int units, bool verifyProvinceIndexParity});
+
+/// Pins for disconnected-tile extraction contribution rows.
+typedef TileContributionDisconnectedPin = ();
+
+enum TileExtractionContributionPin { connectedGrainExcludesCapitalBonus, disconnectedNull }
+
+TileExtractionContributionScenario tileExtractionContributionScenario({required String label, required TileExtractionContributionPin pin, TileContributionConnectedPin? connectedPins, TileMapResult? grainTileMap}) => (label: label, pin: pin, connectedPins: connectedPins, grainTileMap: grainTileMap, refs: null);
+
 /// One row for per-tile extraction contribution scenario tables (Refs #3979).
 typedef TileExtractionContributionScenario = ({
   String label,
@@ -11,20 +21,7 @@ typedef TileExtractionContributionScenario = ({
   TileMapResult? grainTileMap,
   String? refs,
 });
-void runTileExtractionContributionScenario(
-  TileExtractionContributionScenario scenario,
-) {
-  final tileMap = scenario.grainTileMap ?? singleTileMap(Resource.grain);
-  switch (scenario.pin) {
-    case TileExtractionContributionPin.connectedGrainExcludesCapitalBonus:
-      runTileContributionConnectedPin(
-        grainTileMap: tileMap,
-        pins: scenario.connectedPins!,
-      );
-    case TileExtractionContributionPin.disconnectedNull:
-      runTileContributionDisconnectedPin(grainTileMap: tileMap);
-  }
-}
+
 List<TileExtractionContributionScenario> tileExtractionContributionScenarios({
   required TileMapResult grainTileMap,
 }) => [
