+++ packages/colonizethis_economy_test_support/lib/src/tile_extraction_pipeline_scenarios.dart
@@ -1,16 +1,17 @@
 // Table-driven tile extraction pipeline scenarios (Refs #3939 phase 3, #3979).
-import 'tile_extraction_pipeline_expectations.dart';
+
+/// Pins for [resolveTileKeyResourceContext] rows.
+enum ResolveTileKeyResourceContextTarget { validGrain, ironCommodityMapping, invalidKeys }
+
+ResolveTileKeyResourceContextScenario resolveTileKeyResourceContextScenario({required String label, required ResolveTileKeyResourceContextTarget target, String? refs}) => (label: label, target: target, refs: refs);
+
 /// One row in [resolveTileKeyResourceContextScenarios] (Refs #3979).
 typedef ResolveTileKeyResourceContextScenario = ({
   String label,
   ResolveTileKeyResourceContextTarget target,
   String? refs,
 });
-void runResolveTileKeyResourceContextScenario(
-  ResolveTileKeyResourceContextScenario scenario,
-) {
-  runResolveTileKeyResourceContextExpectation(scenario.target);
-}
+
 /// Canonical scenarios for [resolveTileKeyResourceContext].
 // dart format off
 List<ResolveTileKeyResourceContextScenario> resolveTileKeyResourceContextScenarios() => [
@@ -18,11 +19,15 @@ List<ResolveTileKeyResourceContextScenario> resolveTileKeyResourceContextScenari
   resolveTileKeyResourceContextScenario(label: 'maps commodity id via resource.name consistently', target: ResolveTileKeyResourceContextTarget.ironCommodityMapping),
   resolveTileKeyResourceContextScenario(label: 'returns null for invalid or out-of-range keys', target: ResolveTileKeyResourceContextTarget.invalidKeys),
 ];
+
+/// Pins for [resolveTileKeyExtractionContext] rows.
+enum ResolveTileKeyExtractionContextTarget { fromIndex, fallbackGame, missingProvince }
+
+ResolveTileKeyExtractionContextScenario resolveTileKeyExtractionContextScenario({required String label, required ResolveTileKeyExtractionContextTarget target, String? refs}) => (label: label, target: target, refs: refs);
+
 /// One row in [resolveTileKeyExtractionContextScenarios] (Refs #3979).
 typedef ResolveTileKeyExtractionContextScenario = ({String label, ResolveTileKeyExtractionContextTarget target, String? refs});
-void runResolveTileKeyExtractionContextScenario(ResolveTileKeyExtractionContextScenario scenario) {
-  runResolveTileKeyExtractionContextExpectation(scenario.target);
-}
+
 /// Canonical scenarios for [resolveTileKeyExtractionContext].
 List<ResolveTileKeyExtractionContextScenario> resolveTileKeyExtractionContextScenarios() => [
   resolveTileKeyExtractionContextScenario(label: 'resolves province from provincesByFullId index', target: ResolveTileKeyExtractionContextTarget.fromIndex),
