--- packages/colonizethis_economy_test_support/lib/src/worker_economy_expectations.dart
@@ -1,20 +1,0 @@
-// Compact worker labour primitive assertions (Refs #3939 phase 3 slice 33, #3979).
-import 'package:colonizethis_economy/colonizethis_economy.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:colonizethis_test/test.dart';
-import 'worker_economy_scenarios.dart';
-// dart format off
-/// Pins for [effectiveLabourFromIdleCounts] rows.
-typedef IdleLabourPins = ({WorkerIdleCounts idle, int expected});
-void runIdleLabourExpectation(IdleLabourPins pins) {
-  expect(effectiveLabourFromIdleCounts(pins.idle), pins.expected);
-}
-IdleLabourScenario idleLabourScenario({required String label, required IdleLabourPins pins}) => (label: label, pins: pins, refs: null);
-/// Pins for [effectiveLabourForWorkers] rows.
-typedef EffectiveLabourPins = ({WorkerPool workers, Stockpile stockpile, int militaryUnits, int expected});
-void runEffectiveLabourExpectation(EffectiveLabourPins pins) {
-  final labour = effectiveLabourForWorkers(workers: pins.workers, stockpile: pins.stockpile, foodCounts: MilitaryNavyFoodCounts(militaryUnits: pins.militaryUnits));
-  expect(labour, pins.expected);
-}
-EffectiveLabourScenario effectiveLabourScenario({required String label, required EffectiveLabourPins pins}) => (label: label, pins: pins, refs: null);
-// dart format on
